executor/requirements.txt
regex==2021.3.17
pynamodb==5.3.2
boto3==1.26.80
botocore==1.29.80
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
dynamodb-json==1.3
pymongo==3.12.0
hvac==0.11.2
# pytablereader==0.31.3
pytablewriter==0.64.2
importlib-resources==5.10.2
zipp==3.12.0
requests==2.25.1
python-dateutil==2.8.2
pycryptodome==3.15.0
matplotlib==3.6.2
jinja2==3.1.2
modular-sdk
